+++ src/components/InlineDialog/InlineDialog.stories.tsx
@@ -2,21 +2,13 @@
 
 import { Meta, StoryFn } from '@storybook/react';
 import { InlineDialog, InlineDialogProps } from './InlineDialog';
-import { DialogHeader } from '@components/DialogHeader';
 import { DialogBody } from '@components/DialogBody/DialogBody';
-import { Dropdown } from '@components/Dropdown';
-import { DialogFooter } from '@components/DialogFooter';
 import { Modality } from '../../types';
-import { Button, ButtonEmphasis, ButtonStyle } from '@components/Button';
+import { Button } from '@components/Button';
 import { POPPER_STORY_ARGS } from '@components/Popper/types';
-import { action } from '@storybook/addon-actions';
-import { Divider } from '@components/Divider';
-import { Flex } from '@components/Flex';
 import { Box } from '@components/Box';
-import IconDotsVertical16 from '@foundation/Icon/Generated/IconDotsVertical16';
-import { IconExclamationMarkCircle16 } from '@foundation/Icon/Generated';
 import { useRef, useState } from 'react';
-import { Tooltip } from '@components/Tooltip';
+import IconJohanna from '@foundation/Icon/Generated/IconJohanna';
 
 export default {
     title: 'Experimental/InlineDialog',
@@ -27,6 +19,7 @@ export default {
         modality: Modality.Modal,
         offset: [0, 8],
         flip: true,
+        width: 360,
         minWidth: 360,
         maxWidth: 360,
         autoHeight: false,
@@ -74,39 +67,12 @@ const Template: StoryFn<InlineDialogProps> = (args) => {
 
     return (
         <Box className="tw-w-fit">
-            <Button
-                emphasis={ButtonEmphasis.Default}
-                icon={<IconDotsVertical16 />}
-                onClick={() => setIsOpen(!isOpen)}
-                ref={triggerRef}
-            ></Button>
+            <Button icon={<IconJohanna />} onClick={() => setIsOpen(!isOpen)} ref={triggerRef}>
+                Trigger
+            </Button>
             <InlineDialog {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
-                <DialogBody>
-                    <Box className="tw-p-4 tw-text-text">
-                        <Flex justify="start" alignContent="center">
-                            <Tooltip content="Just some Information">
-                                <IconExclamationMarkCircle16 />
-                            </Tooltip>
-                            <p>Information</p>
-                        </Flex>
-                        <Dropdown
-                            enablePortal={false}
-                            onChange={(id) => console.log(id)}
-                            activeItemId="1"
-                            menuBlocks={[
-                                {
-                                    id: 'block1',
-                                    menuItems: [
-                                        { id: '1', title: 'Item 1' },
-                                        { id: '2', title: 'Item 2' },
-                                        { id: '3', title: 'Item 3' },
-                                        { id: '4', title: 'Item 4' },
-                                        { id: '5', title: 'Item 5' },
-                                    ],
-                                },
-                            ]}
-                        />
-
+                <DialogBody padding="comfortable">
+                    <Box className="tw-text-text">
                         <TextExample />
                         <TextExample />
                         <TextExample />
@@ -118,111 +84,7 @@ const Template: StoryFn<InlineDialogProps> = (args) => {
     );
 };
 
-const NoFocusableContentTemplate: StoryFn<InlineDialogProps> = (args) => {
-    const [isOpen, setIsOpen] = useState(false);
-    const triggerRef = useRef<HTMLButtonElement | null>(null);
-
-    return (
-        <Flex justify="start">
-            <Button onClick={action('onClick')} style={ButtonStyle.Default}>
-                Button
-            </Button>
-            <Button onClick={action('onClick')} style={ButtonStyle.Danger}>
-                Button
-            </Button>
-            <Button
-                ref={triggerRef}
-                emphasis={ButtonEmphasis.Default}
-                icon={<IconDotsVertical16 />}
-                onClick={() => setIsOpen(!isOpen)}
-            ></Button>
-            <InlineDialog {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
-                <DialogBody>
-                    <Box className="tw-p-4">
-                        <TextExample />
-                    </Box>
-                </DialogBody>
-            </InlineDialog>
-            <Button onClick={action('onClick')} style={ButtonStyle.Loud}>
-                Button
-            </Button>
-        </Flex>
-    );
-};
-
-const InContext: StoryFn<InlineDialogProps> = (args) => {
-    const [isOpen, setIsOpen] = useState(false);
-    const triggerRef = useRef<HTMLButtonElement | null>(null);
-
-    return (
-        <div>
-            <Flex justify="end">
-                <Button onClick={action('onClick')} emphasis={ButtonEmphasis.Default} style={ButtonStyle.Default}>
-                    Button
-                </Button>
-                <Button ref={triggerRef} onClick={() => setIsOpen(!isOpen)}>
-                    InlineDialog Trigger
-                </Button>
-                <InlineDialog {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
-                    <DialogHeader title="Collaborator Settings" padding="compact" onClose={() => setIsOpen(false)} />
-                    <DialogBody>
-                        <div className="tw-p-2 tw-text-text">
-                            <Dropdown
-                                enablePortal={false}
-                                onChange={(id) => console.log(id)}
-                                activeItemId={'1'}
-                                menuBlocks={[
-                                    {
-                                        id: 'block1',
-                                        menuItems: [
-                                            { id: '1', title: 'Item 1' },
-                                            { id: '2', title: 'Item 2' },
-                                            { id: '3', title: 'Item 3' },
-                                            { id: '4', title: 'Item 4' },
-                                            { id: '5', title: 'Item 5' },
-                                        ],
-                                    },
-                                ]}
-                            />
-                            <p className="tw-flex tw-items-center tw-my-2">Label 2</p>
-
-                            <TextExample />
-                            <TextExample />
-                        </div>
-                    </DialogBody>
-                    <DialogFooter
-                        actionButtons={[
-                            {
-                                children: 'Cancel',
-                                emphasis: ButtonEmphasis.Default,
-                                onClick: () => setIsOpen(false),
-                            },
-                            {
-                                children: 'Confirm',
-                                onClick: () => setIsOpen(false),
-                            },
-                        ]}
-                    ></DialogFooter>
-                </InlineDialog>
-                <Button onClick={action('onClick')} emphasis={ButtonEmphasis.Strong} style={ButtonStyle.Loud}>
-                    Button
-                </Button>
-            </Flex>
-            <Divider />
-            <TextExample />
-            <TextExample />
-            <TextExample />
-            <TextExample />
-            <TextExample />
-            <TextExample />
-            <Button emphasis={ButtonEmphasis.Default} style={ButtonStyle.Default} onClick={action('onClick')}>
-                Focusable and Clickable Button
-            </Button>
-        </div>
-    );
-};
-
-export const WithoutHeaderAndFooter = Template.bind({});
+export const Default = Template.bind({});
 export const AsANonModal = Template.bind({});
 AsANonModal.args = {
     modality: Modality.NonModal,
@@ -245,8 +107,7 @@ export const WithAutoHeight = Template.bind({});
 WithAutoHeight.args = {
     autoHeight: true,
 };
-export const InContextWithDialogHeaderAndDialogFooter = InContext.bind({});
-export const WithNoFocusableContent = NoFocusableContentTemplate.bind({});
+
 export const WithNoRoundedCorners = Template.bind({});
 WithNoRoundedCorners.args = {
     roundedCorners: false,
