+++ src/components/Popover/Popover.stories.tsx
@@ -4,11 +4,11 @@ import { Meta, StoryFn } from '@storybook/react';
 import { Popover } from '@components/Popover/Popover';
 import { Button } from '@components/Button';
 import IconExclamationMarkTriangle16 from '@foundation/Icon/Generated/IconExclamationMarkTriangle16';
-import { Box } from '@components/Box';
 import { Container } from '@components/Container';
 import { POPPER_STORY_ARGS } from '@components/Popper/types';
 import { OverlayProps } from '../../types';
 import { useRef, useState } from 'react';
+import IconJohanna from '@foundation/Icon/Generated/IconJohanna';
 
 export default {
     title: 'Experimental/Popover',
@@ -42,8 +42,8 @@ const Template: StoryFn<OverlayProps> = (args) => {
     const triggerRef = useRef<HTMLButtonElement | null>(null);
     return (
         <div className="tw-h-32 tw-flex tw-justify-center tw-items-center">
-            <Button onClick={() => setIsOpen(!isOpen)} ref={triggerRef}>
-                I am the Trigger
+            <Button icon={<IconJohanna />} onClick={() => setIsOpen(!isOpen)} ref={triggerRef}>
+                Trigger
             </Button>
             <Popover {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
                 <Container padding={8}>
@@ -68,7 +68,7 @@ const TriggeredOnHoverTemplate: StoryFn<OverlayProps> = (args) => {
                 onMouseLeave={() => setIsOpen(false)}
                 onBlur={() => setIsOpen(false)}
             >
-                <span>Information</span> <IconExclamationMarkTriangle16 />
+                <span className="tw-mr-1">Information</span> <IconExclamationMarkTriangle16 />
             </button>
             <Popover {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
                 <Container padding={8}>
@@ -79,36 +79,10 @@ const TriggeredOnHoverTemplate: StoryFn<OverlayProps> = (args) => {
     );
 };
 
-const WithInteractiveContentTemplate: StoryFn<OverlayProps> = (args) => {
-    const [isOpen, setIsOpen] = useState(false);
-    const triggerRef = useRef<HTMLButtonElement | null>(null);
-
-    return (
-        <div className="tw-h-32 tw-flex tw-justify-center tw-items-center">
-            <Button onClick={() => setIsOpen(!isOpen)} ref={triggerRef}>
-                I am the Trigger
-            </Button>
-            <Popover {...args} anchor={triggerRef} open={isOpen} handleClose={() => setIsOpen(false)}>
-                <Box className="tw-p-2 tw-w-[400px]">
-                    <p>
-                        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eos explicabo iure placeat quae quidem
-                        reiciendis sequi, veniam. Cupiditate eos necessitatibus nemo non obcaecati, quod ullam. Beatae
-                        deserunt eum quibusdam ratione.
-                    </p>
-                    <Button>Confirm</Button>
-                </Box>
-            </Popover>
-        </div>
-    );
-};
-
 export const TriggeredOnClick = Template.bind({});
 TriggeredOnClick.args = {};
 export const TriggeredOnHover = TriggeredOnHoverTemplate.bind({});
 TriggeredOnHover.args = {};
-
-export const WithInteractiveContent = WithInteractiveContentTemplate.bind({});
-WithInteractiveContent.args = {};
 
 export const DarkTheme = Template.bind({});
 DarkTheme.args = {
