+++ src/components/Tooltip/Tooltip.stories.tsx
@@ -4,13 +4,6 @@ import { Meta, StoryFn } from '@storybook/react';
 import { POPPER_STORY_ARGS } from '@components/Popper/types';
 import { Tooltip, TooltipProps } from '@components/Tooltip/Tooltip';
 import IconIcon24 from '@foundation/Icon/Generated/IconIcon24';
-import { InlineDialog } from '@components/InlineDialog';
-import { Button } from '@components/Button';
-import { Box } from '@components/Box';
-import { Flex } from '@components/Flex';
-import { Dropdown } from '@components/Dropdown';
-import { DialogBody } from '@components/DialogBody';
-import { useRef, useState } from 'react';
 
 export default {
     title: 'Components/Tooltip',
@@ -56,57 +49,6 @@ const Template: StoryFn<TooltipProps> = (args) => (
     </Tooltip>
 );
 
-const InContextTemplate: StoryFn<TooltipProps> = (args) => {
-    const [isDialogOpen, setIsDialogOpen] = useState(false);
-    const triggerRef = useRef<HTMLButtonElement | null>(null);
-
-    return (
-        <Flex justify="center">
-            <Button ref={triggerRef} onClick={() => setIsDialogOpen(!isDialogOpen)}>
-                Hello
-            </Button>
-            <InlineDialog handleClose={() => setIsDialogOpen(false)} anchor={triggerRef} open={isDialogOpen}>
-                <DialogBody>
-                    <Box className="tw-p-4">
-                        <Box className="tw-mb-2">
-                            <Flex justify="start">
-                                <Tooltip {...args}>
-                                    <IconIcon24 />
-                                </Tooltip>
-                                <p>Information</p>
-                            </Flex>
-                        </Box>
-                        <Dropdown
-                            enablePortal={false}
-                            onChange={(id) => console.log(id)}
-                            activeItemId="1"
-                            menuBlocks={[
-                                {
-                                    id: 'block1',
-                                    menuItems: [
-                                        { id: '1', title: 'Item 1' },
-                                        { id: '2', title: 'Item 2' },
-                                        { id: '3', title: 'Item 3' },
-                                        { id: '4', title: 'Item 4' },
-                                        { id: '5', title: 'Item 5' },
-                                    ],
-                                },
-                            ]}
-                        />
-                        <p className="tw-my-2">
-                            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Consequuntur eveniet harum iste
-                            officia totam unde ut. Aperiam earum laborum nesciunt numquam perferendis ratione, ut.
-                            Asperiores cumque minima nemo officia rerum!
-                        </p>
-
-                        <Button onClick={() => setIsDialogOpen(false)}>Close</Button>
-                    </Box>
-                </DialogBody>
-            </InlineDialog>
-        </Flex>
-    );
-};
-
 export const Default = Template.bind({});
 
 export const OpenByDefault = Template.bind({});
@@ -123,5 +65,3 @@ export const Disabled = Template.bind({});
 Disabled.args = {
     disabled: true,
 };
-
-export const InContext = InContextTemplate.bind({});
